remove some int conversions

+++ src/main/kotlin/day05/main.kt
@@ -8,14 +8,14 @@ import kotlin.math.ceil
 import kotlin.math.floor
 import kotlin.math.pow
 
-fun <T> binaryPartition(sequence: Collection<T>, rightValue: T, leftValue: T): Int {
-    var min = 0
-    var max = 2.0.pow(sequence.size).toInt()
+fun <T> binaryPartition(sequence: Collection<T>, rightValue: T, leftValue: T): Double {
+    var min = 0.0
+    var max = 2.0.pow(sequence.size)
 
     for (char in sequence) {
         when (char) {
-            rightValue -> min = ceil(lerp(min.toDouble(), max.toDouble(), 0.5)).toInt()
-            leftValue -> max = floor(lerp(min.toDouble(), max.toDouble(), 0.5)).toInt()
+            rightValue -> min = ceil(lerp(min, max, 0.5))
+            leftValue -> max = floor(lerp(min, max, 0.5))
         }
     }
 
@@ -25,7 +25,7 @@ fun <T> binaryPartition(sequence: Collection<T>, rightValue: T, leftValue: T): I
 fun getBoardingPassSeatId(pass: String): Int {
     val row = binaryPartition(pass.slice(0..6).toList(), 'B', 'F')
     val col = binaryPartition(pass.slice(7..9).toList(), 'R', 'L')
-    return row * 8 + col
+    return row.toInt() * 8 + col.toInt()
 }
 
 fun main() {
